refactor: architectural renaming (#31)
* refactor: rename components folder files

* refactor: bare metal exoprt in `hooks` folder

* refactor: move all services to client folder

* refactor: move client only utils to `client/utils`
