+++ src/client/services/MediaDatabaseService.ts
@@ -1,6 +1,6 @@
 import { $ } from "@qwik.dev/core";
-import type { MediaFileRecord } from "../types/FileRecord";
-import { IndexedDatabaseService } from "./IndexedDatabaseService";
+import type { MediaFileRecord } from "../../types/FileRecord";
+import { IndexedDatabaseService } from "~/client/services";
 
 export const MediaDatabaseService = $((store: string) =>
   IndexedDatabaseService<MediaFileRecord>(store)
